refactor history offer to use GenericOffer
